refactor: extract MainFrame component from ConversationPage
- Move complex rendering logic and state management from ConversationPage to MainFrame
- Improve code organization by separating concerns
- Reduce file complexity and enhance maintainability
